# docs/assignment-2/openapi.yml
openapi: 3.0.3
info:
  title: StudyCrew REST API
  version: 1.0.0
  description: Versioned API for project, task, meeting and collaboration workflows.
paths:
...
  /api/v1/coordination/availability/:
    get:
      operationId: coordination_availability_get
      parameters:
      - in: query
        name: project
        schema:
          type: string
          format: uuid
        required: true
      - in: query
        name: week_start
        schema:
          type: string
          format: date
        required: true
      tags:
      - coordination
      security:
      - mfaSession: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
    put:
      operationId: coordination_availability_put
      tags:
      - coordination
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AvailabilityWrite'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/AvailabilityWrite'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AvailabilityWrite'
        required: true
      security:
      - mfaSession: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/v1/coordination/calendar/:
    get:
      operationId: coordination_calendar_get
      parameters:
      - in: query
        name: project
        schema:
          type: string
          format: uuid
      - in: query
        name: range_end
        schema:
          type: string
          format: date
        required: true
      - in: query
        name: range_start
        schema:
          type: string
          format: date
        required: true
      tags:
      - coordination
      security:
      - mfaSession: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/v1/coordination/calendar/export/:
    get:
      operationId: coordination_calendar_export_get
      parameters:
      - in: query
        name: project
        schema:
          type: string
          format: uuid
      - in: query
        name: range_end
        schema:
          type: string
          format: date
        required: true
      - in: query
        name: range_start
        schema:
          type: string
          format: date
        required: true
      tags:
      - coordination
      security:
      - mfaSession: []
      responses:
        '200':
          content:
            text/calendar:
              schema:
                type: string
                format: binary
          description: ''
  /api/v1/coordination/claims/:
    get:
      operationId: coordination_claims_get
      parameters:
      - in: query
        name: page
        schema:
          type: integer
          minimum: 1
          default: 1
      - in: query
        name: project
        schema:
          type: string
          format: uuid
        required: true
      - in: query
        name: range_end
        schema:
          type: string
          format: date
      - in: query
        name: range_start
        schema:
          type: string
          format: date
      tags:
      - coordination
      security:
      - mfaSession: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
    post:
      operationId: coordination_claims_post
      tags:
      - coordination
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClaimWrite'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ClaimWrite'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ClaimWrite'
        required: true
      security:
      - mfaSession: []
      responses:
        '201':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/v1/coordination/claims/{id}/respond/:
    post:
      operationId: coordination_claim_response_post
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - coordination
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClaimResponse'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ClaimResponse'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ClaimResponse'
        required: true
      security:
      - mfaSession: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/v1/coordination/claims/{id}/review/:
    post:
      operationId: coordination_claim_review_post
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - coordination
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewWrite'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ReviewWrite'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ReviewWrite'
        required: true
      security:
      - mfaSession: []
      responses:
        '201':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/v1/coordination/claims/{id}/withdraw/:
    post:
      operationId: coordination_claim_withdraw_post
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - coordination
      security:
      - mfaSession: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/v1/coordination/meeting-records/:
    get:
      operationId: coordination_meeting_records_get
      parameters:
      - in: query
        name: page
        schema:
          type: integer
          minimum: 1
          default: 1
      - in: query
        name: project
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - coordination
      security:
      - mfaSession: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/v1/coordination/meeting-records/{id}/:
    get:
      operationId: coordination_meeting_record_get
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - coordination
      security:
      - mfaSession: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
    put:
      operationId: coordination_meeting_record_put
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - coordination
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordWrite'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/RecordWrite'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/RecordWrite'
      security:
      - mfaSession: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/v1/coordination/meeting-records/{id}/actions/:
    post:
      operationId: coordination_meeting_action_post
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - coordination
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActionWrite'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ActionWrite'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ActionWrite'
        required: true
      security:
      - mfaSession: []
      responses:
        '201':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/v1/coordination/meeting-records/{id}/attendance/:
    post:
      operationId: coordination_attendance_post
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - coordination
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttendanceWrite'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/AttendanceWrite'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AttendanceWrite'
        required: true
      security:
      - mfaSession: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/v1/coordination/meeting-records/{id}/confirm/:
    post:
      operationId: coordination_confirm_minutes_post
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - coordination
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmationWrite'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ConfirmationWrite'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ConfirmationWrite'
        required: true
      security:
      - mfaSession: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/v1/coordination/meeting-records/{id}/repeat/:
    post:
      operationId: coordination_repeat_meeting_post
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - coordination
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RepeatWrite'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/RepeatWrite'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/RepeatWrite'
        required: true
      security:
      - mfaSession: []
      responses:
        '201':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/v1/coordination/poll-options/{id}/vote/:
    post:
      operationId: coordination_vote_post
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - coordination
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VoteWrite'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/VoteWrite'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/VoteWrite'
        required: true
      security:
      - mfaSession: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/v1/coordination/polls/:
    get:
      operationId: coordination_polls_get
      parameters:
      - in: query
        name: page
        schema:
          type: integer
          minimum: 1
          default: 1
      - in: query
        name: project
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - coordination
      security:
      - mfaSession: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
    post:
      operationId: coordination_polls_post
      tags:
      - coordination
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PollWrite'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PollWrite'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PollWrite'
        required: true
      security:
      - mfaSession: []
      responses:
        '201':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/v1/coordination/polls/{id}/cancel/:
    post:
      operationId: coordination_cancel_poll_post
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - coordination
      security:
      - mfaSession: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/v1/coordination/polls/{id}/close/:
    post:
      operationId: coordination_close_poll_post
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - coordination
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PollClose'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PollClose'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PollClose'
        required: true
      security:
      - mfaSession: []
      responses:
        '201':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/v1/coordination/subscriptions/:
    get:
      operationId: coordination_subscriptions_get
      tags:
      - coordination
      security:
      - mfaSession: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
    post:
      operationId: coordination_subscriptions_post
      tags:
      - coordination
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriptionWrite'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/SubscriptionWrite'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/SubscriptionWrite'
      security:
      - mfaSession: []
      responses:
        '201':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/v1/coordination/subscriptions/{id}/revoke/:
    post:
      operationId: coordination_subscription_revoke_post
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - coordination
      security:
      - mfaSession: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/v1/coordination/subscriptions/{id}/rotate/:
    post:
      operationId: coordination_subscription_rotate_post
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - coordination
      security:
      - mfaSession: []
      responses:
        '201':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
...
components:
  schemas:
...
    AvailabilityWrite:
      type: object
      description: Reject undeclared and read-only input instead of silently ignoring
        it.
      properties:
        project:
          type: string
          format: uuid
        time_zone:
          type: string
          maxLength: 64
        slots: {}
      required:
      - project
      - slots
      - time_zone
...
    ClaimWrite:
      type: object
      description: Reject undeclared and read-only input instead of silently ignoring
        it.
      properties:
        project:
          type: string
          format: uuid
        title:
          type: string
          maxLength: 160
          minLength: 3
        statement:
          type: string
          maxLength: 6000
          minLength: 10
        artifact_url:
          oneOf:
          - type: string
            format: uri
            default: ''
            maxLength: 2048
          - type: string
            maxLength: 0
        task_id:
          type: string
          format: uuid
          nullable: true
        contributor_ids:
          type: array
          items:
            type: string
            format: uuid
          maxItems: 100
        supersedes_id:
          type: string
          format: uuid
          nullable: true
      required:
      - project
      - statement
      - title
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    mfaSession:
      type: apiKey
      in: cookie
      name: sessionid
      description: Authenticated Django session created only after password and email
        OTP. Unsafe requests also require the csrftoken cookie value in X-CSRFToken.